Update README.md
Added credentials for web page access and MySQL. Note: I tried accessing the web server in Filezilla and MySQL client but failed. I tried the computer lab in Concordia and it worked. Either way, this is not an issue. As long as we're using the correct version of MySQL and PHP, our project will work on any Concordia computer.